--- functions/util/admin.js
@@ -1,7 +1,0 @@
-const admin = require('firebase-admin');
-
-admin.initializeApp();
-
-const db = admin.firestore();
-
-module.exports = { admin, db };
