--- functions/util/config.js
@@ -1,10 +1,0 @@
-module.exports = {
-    apiKey: "AIzaSyA2FPyNJSN_2VN7tGYW3YrtS5JpzfJNqMc",
-    authDomain: "socialape-13440.firebaseapp.com",
-    databaseURL: "https://socialape-13440.firebaseio.com",
-    projectId: "socialape-13440",
-    storageBucket: "socialape-13440.appspot.com",
-    messagingSenderId: "51136174000",
-    appId: "1:51136174000:web:a9ad4dbd9224abcc95a875",
-    measurementId: "G-SSVSJDRZVK"
-  };
